+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/sweep/metrics/TargetedSweepMetrics.java
@@ -19,30 +19,48 @@ package com.palantir.atlasdb.sweep.metrics;
 import java.util.Collection;
 import java.util.Comparator;
 import java.util.Map;
+import java.util.function.Function;
+import java.util.function.Supplier;
 
 import com.codahale.metrics.Gauge;
 import com.google.common.collect.ImmutableMap;
 import com.palantir.atlasdb.AtlasDbMetricNames;
+import com.palantir.atlasdb.cleaner.KeyValueServicePuncherStore;
+import com.palantir.atlasdb.keyvalue.api.KeyValueService;
 import com.palantir.atlasdb.protos.generated.TableMetadataPersistence;
 import com.palantir.atlasdb.sweep.queue.ShardAndStrategy;
 import com.palantir.atlasdb.util.AccumulatingValueMetric;
 import com.palantir.atlasdb.util.AggregateRecomputingMetric;
 import com.palantir.atlasdb.util.CurrentValueMetric;
 import com.palantir.atlasdb.util.MetricsManager;
+import com.palantir.common.time.Clock;
+import com.palantir.common.time.SystemClock;
+import com.palantir.util.CachedComposedSupplier;
+import com.palantir.util.JavaSuppliers;
 
 public final class TargetedSweepMetrics {
     private final Map<TableMetadataPersistence.SweepStrategy, MetricsForStrategy> metricsForStrategyMap;
 
-    private TargetedSweepMetrics(MetricsManager metricsManager, long millis) {
+    private TargetedSweepMetrics(MetricsManager metricsManager,
+                Function<Long, Long> tsToMillis, Clock clock, long millis) {
         metricsForStrategyMap = ImmutableMap.of(
                 TableMetadataPersistence.SweepStrategy.CONSERVATIVE,
-                new MetricsForStrategy(metricsManager, AtlasDbMetricNames.TAG_CONSERVATIVE, millis),
+                new MetricsForStrategy(metricsManager, AtlasDbMetricNames.TAG_CONSERVATIVE, tsToMillis, clock, millis),
                 TableMetadataPersistence.SweepStrategy.THOROUGH,
-                new MetricsForStrategy(metricsManager, AtlasDbMetricNames.TAG_THOROUGH, millis));
+                new MetricsForStrategy(metricsManager, AtlasDbMetricNames.TAG_THOROUGH, tsToMillis, clock, millis));
     }
 
-    public static TargetedSweepMetrics withRecomputingInterval(MetricsManager metricsManager, long millis) {
-        return new TargetedSweepMetrics(metricsManager, millis);
+    public static TargetedSweepMetrics create(MetricsManager metricsManager, KeyValueService kvs, long millis) {
+        return createWithClock(metricsManager, kvs, new SystemClock(), millis);
+    }
+
+    public static TargetedSweepMetrics createWithClock(
+            MetricsManager metricsManager, KeyValueService kvs, Clock clock, long millis) {
+        return new TargetedSweepMetrics(
+                metricsManager,
+                ts -> KeyValueServicePuncherStore.getMillisForTimestamp(kvs, ts),
+                clock,
+                millis);
     }
 
     public void updateEnqueuedWrites(ShardAndStrategy shardStrategy, long writes) {
@@ -77,26 +95,33 @@ public final class TargetedSweepMetrics {
         return currentValues.stream().min(Comparator.naturalOrder()).orElse(-1L);
     }
 
-    private static class MetricsForStrategy {
+    private class MetricsForStrategy {
         private final AccumulatingValueMetric enqueuedWrites = new AccumulatingValueMetric();
         private final AccumulatingValueMetric entriesRead = new AccumulatingValueMetric();
         private final AccumulatingValueMetric tombstonesPut = new AccumulatingValueMetric();
         private final AccumulatingValueMetric abortedWritesDeleted = new AccumulatingValueMetric();
         private final CurrentValueMetric<Long> sweepTimestamp = new CurrentValueMetric<>();
-        private final AggregateRecomputingMetric sweptTimestamp;
+        private final AggregateRecomputingMetric lastSweptTimestamp;
+        private final Gauge<Long> millisSinceOldestEntry;
 
-        private MetricsForStrategy(MetricsManager manager, String strategy, long recomputeMillis) {
-            sweptTimestamp = new AggregateRecomputingMetric(TargetedSweepMetrics::minimum, recomputeMillis);
+        private MetricsForStrategy(MetricsManager manager, String strategy, Function<Long, Long> tsToMillis,
+                Clock wallClock, long recomputeMillis) {
             Map<String, String> tag = ImmutableMap.of(AtlasDbMetricNames.TAG_STRATEGY, strategy);
             register(manager, AtlasDbMetricNames.ENQUEUED_WRITES, enqueuedWrites, tag);
             register(manager, AtlasDbMetricNames.ENTRIES_READ, entriesRead, tag);
             register(manager, AtlasDbMetricNames.TOMBSTONES_PUT, tombstonesPut, tag);
             register(manager, AtlasDbMetricNames.ABORTED_WRITES_DELETED, abortedWritesDeleted, tag);
             register(manager, AtlasDbMetricNames.SWEEP_TS, sweepTimestamp, tag);
-            register(manager, AtlasDbMetricNames.SWEPT_TS, sweptTimestamp, tag);
+
+            lastSweptTimestamp = new AggregateRecomputingMetric(TargetedSweepMetrics::minimum, recomputeMillis);
+            register(manager, AtlasDbMetricNames.LAST_SWEPT_TS, lastSweptTimestamp, tag);
+
+            Supplier<Long> lastSweptMillis = new CachedComposedSupplier<>(tsToMillis, lastSweptTimestamp::getValue);
+            millisSinceOldestEntry = JavaSuppliers.compose(ts -> wallClock.getTimeMillis() - ts, lastSweptMillis)::get;
+            register(manager, AtlasDbMetricNames.LAG_MILLIS, millisSinceOldestEntry, tag);
         }
 
-        private static void register(MetricsManager manager, String name, Gauge<Long> metric, Map<String, String> tag) {
+        private void register(MetricsManager manager, String name, Gauge<Long> metric, Map<String, String> tag) {
             manager.registerMetric(TargetedSweepMetrics.class, name, metric, tag);
         }
 
@@ -121,7 +146,7 @@ public final class TargetedSweepMetrics {
         }
 
         private void updateProgressForShard(int shard, long lastSweptTs) {
-            sweptTimestamp.update(shard, lastSweptTs);
+            lastSweptTimestamp.update(shard, lastSweptTs);
         }
     }
 }